+++ src/HttpKernel/TestKernel.php
@@ -18,6 +18,9 @@ use Symfony\Bundle\MonologBundle\MonologBundle;
 use Symfony\Bundle\SecurityBundle\SecurityBundle;
 use Symfony\Bundle\TwigBundle\TwigBundle;
 use Symfony\Bundle\WebServerBundle\WebServerBundle;
+use Symfony\Cmf\Component\Testing\DependencyInjection\Compiler\TestContainerPass;
+use Symfony\Component\DependencyInjection\Compiler\PassConfig;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\BundleInterface;
 use Symfony\Component\HttpKernel\Kernel;
 
@@ -203,6 +206,15 @@ abstract class TestKernel extends Kernel
 
                 $this->requiredBundles[$class] = new $class();
             }
+        }
+    }
+
+    protected function build(ContainerBuilder $container)
+    {
+        parent::build($container);
+        if (in_array($this->getEnvironment(), ['test', 'phpcr']) && file_exists($this->getKernelDir().'/config/public_services.php')) {
+            $services = require $this->getKernelDir().'/config/public_services.php';
+            $container->addCompilerPass(new TestContainerPass($services), PassConfig::TYPE_OPTIMIZE);
         }
     }
 }