+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 2.1.0 (unreleased)
 ------------------
 
+* **2017-01-18**: Introduce a `TestCompilePass` to make services public, which are needed 
+in i.e. a WebTest.
  * **2017-11-08**: Removed php 5.6 and 7.0 support, removed Symfony 3.0.* and 3.1.* support
  introduce KERNEL_CLASS handling to avoid deprecated KERNEL_DIR, removed usage of `ProcessBuilder`